+++ Public/New-AGMLibGCEInstanceDiscovery.ps1
@@ -329,7 +329,7 @@ Function New-AGMLibGCEInstanceDiscovery ([string]$discoveryfile,[switch]$nobacku
             $ct = Get-Date
             write-host "$ct Running Get-AGMDiskpool"
         }
-        $diskpooldatagrab = Get-AGMDiskpool -filtervalue pooltype=cloud | Select-object name,@{N='srcid';E={$_.cloudcredential.sources.srcid}}
+        $diskpooldatagrab = Get-AGMDiskpool -filtervalue pooltype=cloud  | select-object name,@{N='credentialid';E={$_.cloudcredential.id}},@{N='clusterid';E={$_.cluster.clusterid}}
         if ($textoutput)
         {
             $ct = Get-Date
@@ -346,10 +346,10 @@ Function New-AGMLibGCEInstanceDiscovery ([string]$discoveryfile,[switch]$nobacku
             }
             # we need to learn the srcid
             $credgrab = ($srccredgrab | where-object {($_.credentialid -eq $cred.credentialid) -and ($_.applianceid -eq $cred.applianceid)})
-            if ($credgrab.srcid)
+            if ($credgrab.credentialid)
             {
                 $srcid = $credgrab.srcid
-                $diskpoolgrab = $diskpooldatagrab | where-object {($_.srcid -eq $srcid)}
+                $diskpoolgrab = $diskpooldatagrab | where-object {($_.credentialid -eq $cred.credentialid) -and ($_.clusterid -eq $cred.applianceid)}
                 if ($diskpoolgrab.name)
                 {
                     $poolname = $diskpoolgrab.name